+++ cmd/check/main.go
@@ -7,7 +7,7 @@ import (
 	"os"
 	"sort"
 
-	"github.com/nenad/pact-resource/broker"
+	"github.com/nenad/pact-resource/pkg/broker"
 	"github.com/nenad/pact-resource/pkg/concourse"
 )
