+++ pkg/broker/entities.go
